backend/requirements.txt
Flask
Flask-Cors
python-dotenv
langchain
langchain_ollama
langchain_community
langchain_openai
langchain_groq
langchain_experimental
langchain-core
langchain-huggingface
yfinance
requests
ddgs
pandas
numpy
datetime
pymongo
pydantic
dnspython
bcrypt
matplotlib
langgraph
prophet
seaborn
hyperopt
setuptools
scikit-learn
plotly
langchain_google_genai